Add donator ant artstyle register to artpiece creation
The registration of the artpieces donators wasnt being set because the user had no idea of  what to put in
the input field in the creation view. Now the donator id is a select with all the registered legalEntities.
The artstyle input was changed to a select also. The artpiece controller was updated to fill the artsyles
and the donator id.